code-push

+++ src/app/views/buttons/master/popular-vehicle-list/popular-vehicle-list.component.ts
@@ -3,7 +3,7 @@ import { ApiService } from '../../../../api.service';
 import { Router } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 import { SESSION_STORAGE, StorageService } from 'ngx-webstorage-service'
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-popular-vehicle-list',
   templateUrl: './popular-vehicle-list.component.html',
@@ -242,25 +242,51 @@ mainservice_select(event)
     }
 }
 deleteservice(i) {
-  let data = {
 
-    "Subservice_id": i
+  Swal.fire({
+    title: 'Are you sure?',
+    text: 'If you delete this data it will affect existing user details.',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: 'Yes, delete it!',
+    cancelButtonText: 'No, keep it'
+  }).then((result) => {
+    if (result.value) {
+      let data = {
 
-  }
-  console.log(data);
-  this._api.deletesubservice_list(data).subscribe(
-    (response: any) => {
-      console.log(response);
-      if (response.Code == 200) {
-        alert(response.Message);
-        this.ngOnInit();
+        "Subservice_id": i
+    
       }
-      else {
-        alert("sorry");
-      }
-
+      console.log(data);
+      this._api.deletesubservice_list(data).subscribe(
+        (response: any) => {
+          console.log(response);
+          if (response.Code == 200) {
+            alert("Vehicle Deleted successfully");
+            this.ngOnInit();
+          }
+          else {
+            alert("sorry");
+          }
+    
+        }
+      );
+     
+     
+      // Swal.fire(
+      //   'Deleted!',
+      //   'Your imaginary file has been deleted.',
+      //   'success'
+      // )
+    } else if (result.dismiss === Swal.DismissReason.cancel) {
+      Swal.fire(
+        'Cancelled',
+        'Your data is safe.',
+        'error'
+      )
     }
-  );
+  })
+  
 }
   
 deleteRow_new(dynamic,i) {
